test_case_gen/requirements.txt
flask
python-docx
pandas
openpyxl
requests
beautifulsoup4
openai
